% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestral_states.R
\name{ancestral_states}
\alias{ancestral_states}
\alias{posterior_at_nodes}
\alias{posterior_at_ages}
\title{Functions to calculate the posterior probability of ancestral host
repertoires}
\usage{
posterior_at_nodes(history, nodes, host_tree, state = c(2))

posterior_at_ages(history, ages, tree, host_tree, state = c(2), drop_empty = T)
}
\arguments{
\item{history}{Data frame with posterior samples of interaction histories.}

\item{nodes}{Vector of internal nodes for which to calculate the posterior
probability of \code{state}.}

\item{host_tree}{Host tree}

\item{state}{Which state? Default is 2. For analyses using the 3-state model,
can take the values 1 (potential host) or 2 (actual host).}

\item{ages}{Vector of ages (time points in the past) at which ancestral host
repertoires will be reconstructed.}

\item{tree}{Parasite tree}

\item{drop_empty}{Logical. Remove taxa without any interactions?}
}
\description{
Group of functions to calculate the posterior probabilities of
ancestral host repertoires at internal nodes of the parasite tree or at
specific time points in the past.
}
\section{Functions}{
\itemize{
\item \code{posterior_at_nodes}: Make a matrix with marginal posterior probabilities of interactions at given internal nodes.

\item \code{posterior_at_ages}: Make a list of matrices with posterior probabilities for each interaction of extant lineages at given ages.
}}

\examples{
# read parasite and host tree
data(tree)
data(host_tree)

# read histories sampled during MCMC
data(history)

# calculate the posterior probability of all possible interactions
# at chosen ages (time points in the past)
ages <- c(80,70)
pp_at_ages <- posterior_at_ages(history, ages, tree, host_tree)

# calculate the posterior probability of host repertoires
# at chosen internal nodes of the parasite tree
nodes <- c(129:131)
pp_at_nodes <- posterior_at_nodes(history, nodes, host_tree)

}
