% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_counter.R
\name{events_counter}
\alias{events_counter}
\alias{count_events}
\alias{effective_rate}
\alias{count_gl}
\alias{rate_gl}
\title{Calculate the rate of host-repertoire evolution}
\usage{
count_events(history)

effective_rate(history, tree)

count_gl(history)

rate_gl(history, tree)
}
\arguments{
\item{history}{A data frame containing the character history produced by
RevBayes and read by \code{read_history()}.}

\item{tree}{A phylogenetic tree of the symbiont clade}
}
\description{
Calculate the number of host gains, host losses, and the
effective rate of host repertoire evolution.
}
\section{Functions}{
\itemize{
\item \code{count_events()}: Get the average number of events along the symbiont tree and the highest posterior density interval with 95\% probability (HPD95), based on all MCMC iterations in \code{history}.

\item \code{effective_rate()}: Get the effective rate of evolution, i.e. number
of events per branch unit, along each tree branch. Mean and HP95 are outputted.

\item \code{count_gl()}: Get the average number of host gains and host losses

\item \code{rate_gl()}: Get the average effective rate of host gain and host loss

}}
\examples{
data(tree)
data(history)

# all events
n_events <- count_events(history)
rate <- effective_rate(history,tree)

# gains and losses separately
gl_events <- count_gl(history)
gl_rates <- rate_gl(history, tree)
}
