% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolnets-package.R
\docType{package}
\name{evolnets-package}
\alias{evolnets}
\alias{evolnets-package}
\title{evolnets: Reconstruct Ancestral Networks Inferred In RevBayes}
\description{
RevBayes offers models to infer host-repertoire evolution, but no tools to parse the outputs. This package has the necessary tools to reconstruct ancestral ecological networks based on posterior probabilities of interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/evonetslab/evolnets}
  \item \url{https://evonetslab.github.io/evolnets/}
  \item Report bugs at \url{https://github.com/evonetslab/evolnets/issues}
}

}
\author{
\strong{Maintainer}: Mariana P Braga \email{mpiresbr@gmail.com} (\href{https://orcid.org/0000-0002-1253-2536}{ORCID})

Authors:
\itemize{
  \item Wouter van der Bijl
  \item Michael Landis
}

}
\keyword{internal}
