% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolnets.R
\docType{package}
\name{evolnets}
\alias{evolnets-package}
\alias{evolnets}
\title{evolnets: a package for summarizing inference of network evolution}
\description{
The evolnets package provides three categories of important functions:
rates, ancestral states and samples.
}
\section{Rates}{

These functions are used to calculate effective rates of host-repertoire
evolution: effective_rate( ), count_events( ), rate_gl( ), count_gl( ).
}

\section{Ancestral states}{

These functions are used to calculate the posterior probabilities of
host-parasite interactions at internal nodes of the parasite tree or at
specific time points in the past: posterior_at_nodes( ), posterior_at_ages( ).
}

\section{Samples}{

These functions perform calculations for each sampled host-parasite
network during MCMC: samples_at_ages( ), Q_posterior_at_ages( ),
NODF_posterior_at_ages( ).
}

\section{More information}{

Run vignette("intro_to_evolnets") for an example of how to use the package.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maribraga/evolnets}
  \item \url{https://maribraga.github.io/evolnets/}
  \item Report bugs at \url{https://github.com/maribraga/evolnets/issues}
}

}
\author{
\strong{Maintainer}: Mariana P Braga \email{mpiresbr@gmail.com} (\href{https://orcid.org/0000-0002-1253-2536}{ORCID})

Authors:
\itemize{
  \item Wouter van der Bijl
  \item Michael Landis
}

}
