% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extant_net}
\alias{extant_net}
\title{Extant network}
\format{
\code{matrix}: an incidence matrix with interactions between 66 Pieridae genera (rows) and 33 angiosperm families (cols)
}
\source{
\url{https://onlinelibrary.wiley.com/doi/10.1111/ele.13842}
}
\usage{
extant_net
}
\description{
Extant network
}
\keyword{datasets}
