% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_summary_networks}
\alias{get_summary_networks}
\title{Get summary networks from posterior probabilities of interactions of extant lineages at given
time points in the past}
\usage{
get_summary_networks(
  at_ages,
  threshold,
  ages = NULL,
  weighted = TRUE,
  type = "states",
  state = 2,
  repertoire = "fundamental"
)
}
\arguments{
\item{at_ages}{List of lists, with samples and posterior probabilities for each interaction of
extant lineages at given ages. Usually calculated \code{posterior_at_ages}, see example.}

\item{threshold}{Probability threshold to include an interaction in the network. Interactions
with posterior \code{probability < threshold} will be dropped.}

\item{ages}{Vector of ages (time points in the past) at which samples were retrieved. By default,
uses all ages present in \code{pp_at_ages}.}

\item{weighted}{Logical. Use posterior probabilities as interaction weights?}

\item{type}{One of \code{'states'} or \code{'repertoires'}. If \code{'states'}, will plot the presence of a
state when its posterior probablity is higher than \code{threshold}. If \code{'repertoires'}, will plot
the same but for the given \code{repertoire}. Note that this is also applied to the extant network.}

\item{state}{Which state? Default is 2. For analyses using the 3-state model, choose \code{1}, \code{2} or
both using \code{c(1, 2)} (where 1 is a potential host and 2 an actual host). Only used if \code{type} is
\code{'states'}. Note that this is also applied to the extant network.}

\item{repertoire}{Either the \code{'realized'} repertoire which is defined as state 2, or the
\code{'fundamental'} repertoire (default) which is defined as having any state (usually 1 or 2), and
in the 3-state model includes both actual and potential hosts. Note that this is also applied
to the extant network.}
}
\value{
A list of incidence matrices (summary network) for each time slice in \code{ages}.
}
\description{
Get summary networks from posterior probabilities of interactions of extant lineages at given
time points in the past
}
\examples{
data(tree)
data(host_tree)
data(history)

ages <- c(60, 50, 40, 0)
at_ages <- posterior_at_ages(history, ages, tree, host_tree)
weighted_net_50 <- get_summary_networks(at_ages, threshold = 0.5, weighted = TRUE)
binary_net_90 <- get_summary_networks(at_ages, threshold = 0.9, weighted = FALSE)
}
