% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{host_tree}
\alias{host_tree}
\title{Host plants tree}
\format{
\code{host_tree}: a phylogenetic tree of angiosperm families with 50 tips.
}
\source{
Braga et al. 2020 BioRxiv
}
\usage{
host_tree
}
\description{
Host plants tree
}
\keyword{datasets}
