% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NODF_posterior_at_ages.R
\name{index_at_ages}
\alias{index_at_ages}
\title{Posterior distribution of indices of network structure across ancestral networks}
\usage{
index_at_ages(samples_at_ages, ages, index, null = 100, seed = NULL)
}
\arguments{
\item{samples_at_ages}{List of ancestral networks sampled across MCMC at
given ages.}

\item{ages}{Vector of ages (time points in the past) of ancestral networks.}

\item{index}{Index to be calculated for each ancestral network. "NODF" to
calculate nestedness and "Q" to calculate modularity.}

\item{null}{Number of null networks to generate to calculate the z-score.}

\item{seed}{Seed passed to \code{stats::simulate} to generate null networks.
Default to NULL.}
}
\value{
A tibble of NODF z-scores and p-values across samples and ages.
}
\description{
Calculate z-scores for nestedness (NODF) or modularity (Q) for each MCMC sample at
time points in the past based on null networks where all interactions have
the same probability. By calculating z-scores, we can compare ancestral
networks at different ages.
}
\examples{
# calculate posterior distribution of nestedness
#Nz <- index_at_ages(samples_at_ages, ages, index = "NODF", null = 100, seed = NULL)

# calculate posterior distribution of modularity
#Qz <- index_at_ages(samples_at_ages, ages, index = "Q", null = 100, seed = NULL)
}
