% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_across_ages}
\alias{modules_across_ages}
\title{Evolution of modules: find and match modules across time slices}
\usage{
modules_across_ages(summary_networks, tree)
}
\arguments{
\item{summary_networks}{List of reconstructed summary networks for each age.}

\item{tree}{The phylogeny, a \code{phylo} object.}
}
\value{
A data frame with module membership for each node of summary networks
at each age.
}
\description{
Evolution of modules: find and match modules across time slices
}
\examples{

}
