% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modularity_alternatives.R
\name{mycomputeModules}
\alias{mycomputeModules}
\title{Modified version of bipartite's computeModules function}
\usage{
mycomputeModules(
  web,
  method = "Beckett",
  steps = 1e+06,
  tolerance = 1e-10,
  forceLPA = FALSE
)
}
\arguments{
\item{web}{Incidence matrix}

\item{method}{Becket}

\item{steps}{steps}

\item{tolerance}{tolerance}

\item{forceLPA}{FALSE}
}
\value{
Modularity Q
}
\description{
Modified version of bipartite's computeModules function
}
\examples{
\dontrun{
  data_path <- system.file("extdata", package = "evolnets")
  extant_net <- read.csv(paste0(data_path,"/interaction_matrix_pieridae.csv"), row.names = 1)

  mod <- mycomputeModules(extant_net)
}
}
