% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_extant_matrix}
\alias{plot_extant_matrix}
\title{Plot a network with modules as an adjacency matrix}
\usage{
plot_extant_matrix(
  net,
  modules = NULL,
  module_order = NULL,
  parasite_order = NULL,
  host_order = NULL,
  state_alpha = c(0.3, 1)
)
}
\arguments{
\item{net}{An adjacency matrix for a bipartite network. Parasites should be
the rows, hosts should be columns. If all values are 0 or 1 an unweighted
network is represented, otherwise a weighted network is assumed.}

\item{modules}{A \code{moduleWeb} or a \code{data.frame} object defining the models in
the network. If left \code{NULL} the modules are automatically calculated. If a
\code{data.frame} is passed, it must contain three columns: $name with taxon
names, $module with the module the taxon was assigned to, and $type which
defines if the taxon is a "host" or a "symbiont".}

\item{module_order}{A character vector giving the order that modules should
be plotted. Should contain each module only once.}

\item{parasite_order}{A character vector giving the order the parasite should
be listed in. Should contain each parasite only once, and include the row
names of \code{net}.}

\item{host_order}{A character vector giving the order the hosts should be
listed in. Should contain each host only once, and include the column names
of \code{net}.}

\item{state_alpha}{A numeric vector of length 2. Gives the alpha
(transparency) values for the interaction type in the three-state model}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a network with modules as an adjacency matrix
}
\examples{
\dontrun{
  # The slow portion of this function is the calculation of the modules.
  plot_extant_matrix(extant_net)

  # Change our network to a weighted one:
  extant_net_weighted <- extant_net
  extant_net_weighted[extant_net == 1] <- runif(sum(extant_net))
  plot_extant_matrix(extant_net_weighted)
}
}
