% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_at_ages.R
\name{plot_index_at_ages}
\alias{plot_index_at_ages}
\title{Plot network structure indices across ancestral summary, sampled and extant networks}
\usage{
plot_index_at_ages(
  nodf_sampled,
  q_sampled = NULL,
  nodf_summary = NULL,
  q_summary = NULL,
  col_sampled = "grey40",
  col_summary = "black"
)
}
\arguments{
\item{nodf_sampled}{Output of \code{index_at_ages_samples} when index = "NODF".}

\item{q_sampled}{Output of \code{index_at_ages_samples} when index = "Q".}

\item{nodf_summary}{Output of \code{index_at_ages_summary} when index = "NODF".}

\item{q_summary}{Output of \code{index_at_ages_summary} when index = "Q".}

\item{col_sampled}{Color used to represent values from sampled networks.}

\item{col_summary}{Color used to represent values from summary networks.}
}
\value{
A plot of z-scores over time. Violins show the posterior distribution of z-scores of sampled networks, with dots and lines showing the mean values. Z-scores of summary networks and extant network are shown as dots and line. Use different colors to differentiate values from sampled and summary networks.
}
\description{
Plot z-scores for nestedness (NODF) and/or modularity (Q) for sampled and summary networks at time points in
the past, calculated by \code{index_at_ages_samples} and \code{index_at_ages_summary}.
}
\examples{
data(tree)
data(host_tree)
data(history)

ages <- c(60, 50, 40, 0)
at_ages <- posterior_at_ages(history, ages, tree, host_tree)

# summary networks
summary_networks <- get_summary_networks(at_ages, threshold = 0.5, weighted = TRUE)
Nz_sum <- index_at_ages_summary(summary_networks, index = "NODF", nnull = 10)

# sampled networks
sampled_networks <- get_sampled_networks(at_ages)
Nz_sam <- index_at_ages_samples(sampled_networks, index = "NODF", nnull = 10)

#plot
plot_index_at_ages(nodf_sampled = Nz_sam, nodf_summary = Nz_sum)
}
