% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_matrix_phylo}
\alias{plot_matrix_phylo}
\title{Plot a network with modules as an adjacency matrix, with aligned phylogenies.}
\usage{
plot_matrix_phylo(
  net,
  samples_at_nodes,
  tree,
  host_tree,
  type = "states",
  state = 2,
  repertoire = "fundamental",
  modules = NULL,
  module_order = NULL,
  threshold = 0.9,
  point_size = 3,
  dodge_width = 0.025,
  colors = NULL,
  ladderize = FALSE
)
}
\arguments{
\item{net}{An adjacency matrix for a bipartite network. This should be the extant network.
Parasites should be the rows, hosts should be columns. If all values are 0 or 1 an binary
network is represented, otherwise a weighted network is assumed.}

\item{samples_at_nodes}{A list of length 2, output from \code{posterior_at_nodes()}.}

\item{tree}{The phylogeny of the symbiont clade (e.g. parasites, herbivores). Object of class \code{phylo}.}

\item{host_tree}{The phylogeny belonging to the hosts. Object of class \code{phylo}.

See the examples on how to change the color scale.}

\item{type}{One of \code{'states'} or \code{'repertoires'}. If \code{'states'}, will plot the presence of a
state when its posterior probability is higher than \code{threshold}. If \code{'repertoires'}, will plot
the same but for the given \code{repertoire}.}

\item{state}{Which state? Default is 2. For analyses using the 3-state model, choose \code{1} or \code{2}
(where 1 is a potential host and 2 an actual host). Only used if \code{type} is \code{'states'}.}

\item{repertoire}{Either the \code{'realized'} repertoire which is defined as state 2, or the
\code{'fundamental'} repertoire (default) which is defined as having any state (usually 1 or 2), and
in the 3-state model includes both actual and potential hosts.}

\item{modules}{A \code{moduleWeb} or a \code{data.frame} object defining the modules in the network.
If a \code{data.frame} is passed, it must contain three columns:
$name with taxon names,
$module with the module the taxon was assigned to, and
$type which defines if the taxon is a "host" or a "symbiont".}

\item{module_order}{A character vector giving the order that modules should be plotted. Should contain
each module only once.}

\item{threshold}{The posterior probability above which the ancestral states should be shown.
Defaults to 90\% (\code{0.9}). Numeric vector of length 1.}

\item{point_size}{How large the ancestral state points should be, default at 3. Play with this
and \code{dodge_width} to get a pleasing result. Numeric vector of length 1.}

\item{dodge_width}{How far the points should be pushed apart, when there is multiple states at
a single node, default at 0.025. Play with this and \code{point_size} to get a pleasing result.
Numeric vector of length 1.}

\item{colors}{Override the default colors. Should be a character vector with as many color values
as there are modules.}

\item{ladderize}{Logical. Whether to ladderize the tree. Default to FALSE.

The ancestral states are automatically colored by module. To change what colors are used, you
can add color scales to the resulting \code{ggplot}, e.g. \code{ggplot2::scale_color_manual()}.}
}
\value{
An assembly of plots, of class \code{patchwork}.
}
\description{
Plot a network with modules as an adjacency matrix, with aligned phylogenies.
}
\examples{
\dontrun{
  san <- posterior_at_nodes(history, tree, host_tree, 66 + 1:65)
  plot_matrix_phylo(extant_net, san, tree, host_tree)
  # manual_colors
  plot_matrix_phylo(extant_net, san, tree, host_tree, colors = rainbow(20))
}
}
