% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network_at_age}
\alias{plot_network_at_age}
\title{Plot one ancestral network with module information at a given time}
\usage{
plot_network_at_age(subtree, tip_data, tgraph, module_levels, palette)
}
\arguments{
\item{subtree}{a \code{phylo} object of the original tree sliced at a given time in the past.}

\item{tip_data}{a \code{data.frame} containing the module information for each tip in the subtree.}

\item{tgraph}{a \code{tbl_graph} containing the nodes and edges of the ancestral network and the module information for each node.}

\item{module_levels}{Order in which the modules should be organized. Affects which color each module will be assigned.}

\item{palette}{Color palette used to plot module information.}
}
\value{
An assembly of plots, of class \code{patchwork}.
}
\description{
Plot one ancestral network with module information at a given time
}
\examples{
\dontrun{
}
}
