% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_at_ages.R
\name{posterior_at_ages}
\alias{posterior_at_ages}
\title{Calculate the posterior probability of ancestral host repertoires at time points in the past (ages)}
\usage{
posterior_at_ages(
  history,
  ages,
  tree,
  host_tree,
  extant_prob = FALSE,
  state = 2,
  drop_empty = TRUE
)
}
\arguments{
\item{history}{Data frame with posterior samples of interaction histories. Output from \code{read_history()}.}

\item{ages}{Vector of ages (time points in the past) at which samples will be retrieved. The
present (age = 0) must be included.}

\item{tree}{Symbiont tree.}

\item{host_tree}{Host tree.}

\item{extant_prob}{Should posterior probabilities be calculated for extant network? Default to
FALSE. TRUE only makes sense if interactions in the extant network were also inferred. When
FALSE, only the first MCMC sample will be retrieved at age = 0.}

\item{state}{Probabilities will be calculated only for this state. Default is 2. For analyses
using the 3-state model, can take the values 1 (potential host) or 2 (actual host). For
analyses using the 2-state model, state = 1 is ignored.}

\item{drop_empty}{Logical. Remove taxa without any interactions?}
}
\value{
A list of three lists, containing:
\itemize{
\item{"\code{samples}"}{ Arrays of samples x nodes x hosts, containing the state of each sample.}
\item{"\code{post_states}"}{ Arrays of nodes x hosts x state containing the posterior probability
for each state.}
\item{"\code{post_repertoires}"}{ Arrays of nodes x hosts x repertoire containing the posterior
probability for 1) the \code{"realized"} repertoire which is defined as state 2, and 2) the
\code{"fundamental"} repertoire which is defined as having any state (usually 1 or 2).}
}
Each array in these lists is for one of the \code{ages}. The number of samples is the number of
iterations in \code{history}.
}
\description{
Calculate the posterior probability of ancestral host repertoires at time points in the past (ages)
}
\examples{
# read symbiont and host tree
data(tree)
data(host_tree)

# read histories sampled during MCMC
data(history)

# get samples at ages
ages <- c(80,70,0)
at_ages <- posterior_at_ages(history, ages, tree, host_tree)
samples_at_ages <- at_ages$samples

# get posterior probabilities at ages
pp_at_ages <- at_ages$post_states
}
