% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_at_nodes.R
\name{posterior_at_nodes}
\alias{posterior_at_nodes}
\title{Functions to calculate the posterior probability of ancestral host
repertoires at internal nodes}
\usage{
posterior_at_nodes(history, nodes, host_tree, state = c(2))
}
\arguments{
\item{history}{Data frame with posterior samples of interaction histories.}

\item{nodes}{Vector of internal nodes for which to calculate the posterior
probability of \code{state}.}

\item{host_tree}{Host tree}

\item{state}{Which state? Default is 2. For analyses using the 3-state model,
can take the values 1 (potential host) or 2 (actual host).}
}
\value{
A matrix with marginal posterior probabilities of interactions at given internal nodes.
}
\description{
Group of functions to calculate the posterior probabilities of
ancestral host repertoires at internal nodes of the parasite tree.
}
\examples{
# read parasite and host tree
data(tree)
data(host_tree)

# read histories sampled during MCMC
data(history)

# calculate the posterior probability of host repertoires
# at chosen internal nodes of the parasite tree
nodes <- c(129:131)
pp_at_nodes <- posterior_at_nodes(history, nodes, host_tree)
}
