% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tree_from_revbayes}
\alias{read_tree_from_revbayes}
\title{Read symbiont tree file produced by RevBayes containing node indices}
\usage{
read_tree_from_revbayes(path_to_tree)
}
\arguments{
\item{path_to_tree}{String path to the .txt file exported from RevBayes}
}
\value{
A \code{phylo} object with node names given by RevBayes, which are important to place the inferred ancestral states in the symbiont tree.
}
\description{
Read symbiont tree file produced by RevBayes containing node indices
}
\examples{
\dontrun{tree <- read_tree_from_revbayes("symbiont_tree.tre")}
}
