% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_modules.R
\name{validate_modules}
\alias{validate_modules}
\title{Validate modules from summary networks}
\usage{
validate_modules(
  samples_at_ages,
  ages,
  modules_across_ages,
  include_all = FALSE,
  seed = NULL,
  edge_list = TRUE
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
Functions to validate modules from summary networks by calculating the frequency with
which pairs of nodes are placed in the same module in networks sampled across MCMC
}
