% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_counter.R
\name{events_counter}
\alias{events_counter}
\alias{count_events}
\alias{effective_rate}
\alias{count_gl}
\alias{rate_gl}
\title{Calculate the rate of host-repertoire evolution}
\usage{
count_events(history)

effective_rate(history, tree)

count_gl(history)

rate_gl(history, tree)
}
\arguments{
\item{history}{A data frame containing the character history produced by RevBayes}

\item{tree}{A phylogenetic tree of the parasite clade}
}
\description{
Calculate the number of host gains, host losses, and the effective rate of host repertoire evolution
}
\section{Functions}{
\itemize{
\item \code{count_events}: Get the total number of events along the parasite tree

\item \code{effective_rate}: Get the effective rate of evolution, i.e. number of events per million years, along each tree branch

\item \code{count_gl}: Get the number of host gains and host losses

\item \code{rate_gl}: Get the effective rate of host gain and host loss
}}

\examples{
# read parasite tree
data(tree)

# read hist
data(history)

# remove burnin and thin samples
it_seq <- seq(10000,100000,1000)
history <- dplyr::filter(history, iteration \%in\% it_seq)

# effective rate
n_events <- count_events(history)
rate <- effective_rate(history,tree)

# gains and losses
gl <- count_gl(history)
rgl <- rate_gl(history, tree)

}
