% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolnets.R
\docType{package}
\name{evolnets}
\alias{evolnets}
\title{evolnets: a package for parsing outputs from inference of host-repertoire evolution}
\description{
The evolnets package provides three categories of important functions:
rates, ancestral states and samples.
}
\section{Rates functions}{

The rates functions are used to calculate effective rates of host-repertoire
evolution.
}

\section{Ancestral states}{

The ancestral states functions are used to calculate the posterior probabilities
of host-parasite interactions at internal nodes of the parasite tree or at
specific time points in the past.
}

\section{Samples}{

The samples functions perform calculation for each sampled host-parasite
network during MCMC.
}

