% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_networks.R
\name{get_summary_network}
\alias{get_summary_network}
\title{Get summary networks from posterior probabilities of interactions of extant
lineages at given time points in the past}
\usage{
get_summary_network(pp_at_ages, ages, pt, weighted = TRUE)
}
\arguments{
\item{pp_at_ages}{List of matrices with posterior probabilities for each
interaction of extant lineages at given ages}

\item{ages}{Vector of ages (time points in the past) at which samples were retrieved.}

\item{pt}{Probability threshold to include an interaction in the network.
Interactions with posterior probability < pt will be dropped.}

\item{weighted}{Logical. Use posterior probabilities as interaction weights?}
}
\value{
A list of incidence matrices (summary network) for each time slice in \code{ages}.
}
\description{
Get summary networks from posterior probabilities of interactions of extant
lineages at given time points in the past
}
\examples{
data(tree)
data(host_tree)
data(history)

ages <- c(60,50,40)
pp_at_ages <- posterior_at_ages(history, ages, tree, host_tree)[[2]]
weighted_net_50 <- get_summary_network(pp_at_ages, pt = 0.5, weighted = TRUE)
binary_net_90 <- get_summary_network(pp_at_ages, pt = 0.9, weighted = FALSE)
}
