% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_at_ages.R
\name{index_at_ages_samples}
\alias{index_at_ages_samples}
\title{Posterior distribution of network structure indices across ancestral networks}
\usage{
index_at_ages_samples(samples_at_ages, index, ages = NULL, nnull = 100)
}
\arguments{
\item{index}{Index to be calculated for each ancestral network. "NODF" to calculate nestedness or
"Q" to calculate modularity.}

\item{ages}{Vector of ages (time points in the past) of ancestral networks. By default, uses all
ages present in \code{sampled_networks}.}

\item{nnull}{Number of null networks to generate to calculate the z-score. Default is 100.}

\item{sampled_networks}{List of ancestral networks sampled across MCMC at given ages. Usually from
the output of \code{get_sampled_networks}, see example.}
}
\value{
A data.frame of z-scores and p-values across samples and ages.
}
\description{
Calculate z-scores for nestedness (NODF) or modularity (Q) for each MCMC sample at time points in
the past based on null networks where all interactions have the same probability. By calculating
z-scores, we can compare ancestral networks at different ages.
}
\examples{
data(tree)
data(host_tree)
data(history)

ages <- c(60,50,40,0)
samples_at_ages <- posterior_at_ages(history, ages, tree, host_tree)
sampled_networks <- get_sampled_networks(samples_at_ages)

# calculate posterior distribution of nestedness
Nz <- index_at_ages(sampled_networks, index = "NODF")

#  calculate posterior distribution of modularity (SLOW!)
# Qz <- index_at_ages(sampled_networks, index = "Q")
}
