% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_from_samples}
\alias{modules_from_samples}
\title{Find modules in networks sampled across MCMC at specific time slices}
\usage{
modules_from_samples(samples_at_ages)
}
\arguments{
\item{samples_at_ages}{List of sampled networks at time slices produces by calling posterior_at_ages().}
}
\value{
Data frame with module membership information for each sampled network at each time slice.
}
\description{
Find modules in networks sampled across MCMC at specific time slices
}
\examples{
\dontrun{
  mod_samples <- modules_from_samples(samples_at_ages, ages)
}
}
