% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_from_samples.R
\name{modules_from_samples}
\alias{modules_from_samples}
\title{Find modules in networks sampled across MCMC at specific time slices}
\usage{
modules_from_samples(samples_at_ages, ages, seed = NULL)
}
\arguments{
\item{samples_at_ages}{List of sampled networks at time slices produces by calling posterior_at_ages().}

\item{ages}{Vector of ages (time points in the past) at which samples will be
retrieved.}

\item{seed}{Seed set before calling the modularity function.}
}
\value{
Data frame with module membership information
}
\description{
Find modules in networks sampled across MCMC at specific time slices
}
\examples{

}
