% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mycomputeModules}
\alias{mycomputeModules}
\title{Modified version of bipartite's computer_module function}
\usage{
mycomputeModules(
  web,
  method = "Beckett",
  steps = 1e+06,
  tolerance = 1e-10,
  forceLPA = FALSE
)
}
\arguments{
\item{web}{Incidence matrix}

\item{method}{Becket}

\item{steps}{steps}

\item{tolerance}{tolerance}

\item{forceLPA}{FALSE}
}
\value{
Modularity Q
}
\description{
Modified version of bipartite's computer_module function
}
\examples{
\dontrun{
  extant_net <- data(extant_net)
  mod <- mycomputeModules(extant_net)
}
}
