% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_history}
\alias{read_history}
\title{Read history.txt file produced by RevBayes}
\usage{
read_history(path_to_hist, burnin = 0.1)
}
\arguments{
\item{path_to_hist}{String path to the .txt file exported from RevBayes}

\item{burnin}{Fraction of iterations to be removed as burnin. Default to 10\%.}
}
\value{
A data frame
}
\description{
Read history.txt file containing the posterior samples of
interaction histories
}
\examples{
# read history file
\dontrun{history <- read_history("/path/history.txt", burnin = 0.2)}
}
