% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Pieridae tree}
\format{
\code{tree}: a phylogenetic tree of Pieridae genera with 66 tips.
}
\source{
Braga et al. 2020 BioRxiv
}
\usage{
tree
}
\description{
Pieridae tree
}
\keyword{datasets}
